Validate client's TLS certificates, when CA provided
If a root CA is passed, we set RequireAndVerifyClientCert on the tls
config, which ensures that a client certificate should be requested
during the handshake, and that at least one valid certificate is
required to be sent by the client.

If no root CA is passed, no client certificate will be requested during
the handshake, and if any certificates are sent they will not be
verified.

Fixes #76